backend/protobackend/requirements.txt
asgiref==3.2.10
Django==3.1.1
django-braces==1.14.0
django-cors-headers==3.5.0
django-extensions==3.0.9
djangorestframework==3.11.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
